… Sources/SwiftSyntax/generated/SyntaxRewriter.swift
… Sources/SwiftSyntax/generated/SyntaxVisitor.swift