## Description

This PR optimizes swift code to accommodate slowness of Type Checker.

## Context

The following two methods are currently slow to compile in `swift-syntax` package:

1. `TokenKind.fromRaw`
2. `SyntaxVisitor.visitationFunc`
3. `SyntaxRewriter.visitationFunc`
4. `SyntaxUtils.init?(
    combining syntax1: some UnexpectedNodesCombinable,
    _ syntax2: some UnexpectedNodesCombinable,
    _ syntax3: some UnexpectedNodesCombinable,
    _ syntax4: some UnexpectedNodesCombinable,
    arena: __shared SyntaxArena
  )`

Optimization in this PR includes the following:

1. change in `SyntaxVisitorFile` template in code-generation package to include the name of the type in `visit....()` methods
2. change in `SyntaxRewriterFile` template in code-generation package to include the name of the type in `visit...()` methods
3. change in `TokenKindFile` template in code-generation package to add a private extension of `RawTokenKind`
4. manual change in `SyntaxUtils.init`

## Measurements

I have measured the performance of the following command:
```bash
time xcodebuild -scheme swift-syntax-Package ONLY_ACTIVE_ARCH=YES -destination 'platform=macos' clean build > /dev/null
```

### Comparison of `time` output

Before|After
:-:|:-:
2.24s user 0.61s system 20% cpu 13.775 total|2.21s user 0.57s system 20% cpu 13.510 total
2.25s user 0.59s system 20% cpu 13.994 total|2.25s user 0.60s system 20% cpu 13.858 total
2.23s user 0.59s system 20% cpu 13.989 total|2.21s user 0.58s system 20% cpu 13.556 total
2.24s user 0.58s system 20% cpu 13.837 total|2.18s user 0.57s system 20% cpu 13.745 total
2.26s user 0.59s system 20% cpu 13.812 total|2.21s user 0.60s system 20% cpu 13.686 total
2.15s user 0.63s system 20% cpu 13.721 total|2.10s user 0.60s system 19% cpu 13.909 total

### Comparison of `build` operation duration in Xcode

Before (seconds)|After (seconds)
:-:|:-:
12.225|11.923 
12.068|11.745 
12.087|11.872 
12.358|11.929 
12.323|12.199 
12.307|12.178 

Overall decrease in compilation (on average): **2.07%**

### Comparison of SwiftCompilationTimingParser output

Before (ms) | After (ms)
:-:|:-:
SUM 6,128.71|SUM 5,415.44
AVERAGE 11.476985019|AVERAGE 9.900255941
MIN 4|MIN 4
MAX 485.9|MAX 353.48
COUNT 535|COUNT 548

Overall, decrease in compilation of slow symbols: **~12%**

### More Details

I have used https://github.com/qonto/SwiftCompilationTimingParser to measure slow compiling code. The following were the top candidates for optimization:

location|symbol|ms
-- | -- | --
Sources\/SwiftSyntax\/generated\/SyntaxVisitor.swift:3438:16 | instance method visitationFunc(for:) | 485.9
Sources\/SwiftSyntax\/generated\/SyntaxRewriter.swift:2113:16 | instance method visitationFunc(for:) | 315.72
Sources\/SwiftSyntax\/generated\/TokenKind.swift:730:22 | static method fromRaw(kind:text:) | 113.12
Sources\/SwiftSyntax\/generated\/ChildNameForKeyPath.swift:18:13 | global function childName(_:) | 371.07
Sources\/SwiftSyntax\/generated\/SyntaxEnum.swift:306:8 | instance method as(_:) | 102.14


Out of these, only the mentioned 3 were optimizable, whereas `SyntaxEnum` and `ChildNameForKeyPath` are slow due to use of `AnyKeyPath` and keypath operator `\`, which I was not able to optimize.
After code optimization, the modified methods got the following timing:

location|symbol|ms|improvement
-- | -- | -- | --
SyntaxVisitor.swift:3438:16 | instance method visitationFunc(for:)|42.98 | **91.16%**
SyntaxRewriter.swift:2113:16 | instance method visitationFunc(for:)|33.33 | **89.44%**
TokenKind.swift:739:22 | static method fromRaw(kind:text:)|33.83 | **70.09%**


Follow-up of #2308