@art-divin 2023-11-05
Hello, dear Swift team 👋🏻 

I wish you all a great week ahead, and I hope you'll have a chance to review this PR. I have invested more than 1 month into figuring out how to speed up type checker for these use cases, and I could not speed it up more than this, next step would be (for me personally) to move to Swift repository and fix issues inside of the type checker itself.

Thank you for your work 🤝