@ahoppen 2024-02-13
Hi @art-divin,

We just discussed your PR and decided that the added complexity in lines of code is not worth the marginal compile time improvement. I really appreciate the effort you put into this PR and respect the dedication with which you continued measuring the compile times.